shmem: implement used and dirty pages tracking
To track anon shared vma pages usage we create a bitmap for it.
Each bit in this bitmap corresponds to particular page in vma. Bit with
value of 1 in this map states that page is used. Bit 0 states that the
page wasn't ever used by any dumpee process.
Page is considered used if it has at least one PME_PRESENT or PME_SWAP
bit set in any dumpee process pagemap.
Pages usage tracking allows us not to dump unused pages at all.

To track anon shared vma pages dirtiness we create a bitmap for it.
Each bit in this bitmap corresponds to particular page in vma. Bit with
value of 1 in this map states that page is dirty. Bit 0 states that the
page is clean.
Page is considered dirty if it has at least one PME_SOFT_DIRTY bit set in
any dumpee process pagemap.
Dirty pages tracking allows not to dump used pages on incremental dumps
if pages contents haven't changed.

Signed-off-by: Fyodor Bocharov <fbocharov@yandex.ru>
Signed-off-by: Eugene Batalov <eabatalov89@gmail.com>